Replaced TCPStream for Audio with tokio::mspc - Generated QuicNetworkPackets are now passed via tokio::mspc to main thread which should eliminate a bunch of overhead - Audio threads can now be stopped via moka::sync::Cache instead of moka::future::Cache - Removed some unecessary threading in quic server
@todo()! there's an outstanding issue where the clients will "fall
behind" the server. Current attempt is to force a flush of the quic
stream to ensure that the server / client ack's the data. This seems to
help, but under many receive loads (tested locally, which is probably
contributing to the issue due to the threading), the CPU load on the
server slowly ramps up and the inbound stream eventually slows.
Need to test this with an actual audio stream to see if the impact is
meaningful, of if it's just a tracing::info!() write output issue.